Code cleanup
Shifted additional int-based calculations to long-based calculations

Cleaned up JUnit all test cases (100% pass)

+++ src/com/dndsuite/core/json/parsers/subevents/AttackRoll.java
@@ -27,7 +27,7 @@ public class AttackRoll extends Subevent implements Calculation {
 
 		// Load critical hit threshold (default is 20)
 		if (json.containsKey("critical_threshold")) {
-			criticalThreshold = (int) json.get("critical_threshold");
+			criticalThreshold = (long) json.get("critical_threshold");
 		} else {
 			criticalThreshold = 20;
 		}
